delete some mediumint type and add function that the mo-tester support pass the para through command

+++ cases/operator/is_operator.sql
@@ -218,7 +218,7 @@ where id in (select id from t2 as x1 where (t2.cur_date is null));
 -- @label:bvt
 DROP TABLE IF EXISTS t1,t2;
 CREATE TABLE `t1` (
-  `numeropost` mediumint(8) unsigned NOT NULL auto_increment,
+  `numeropost` int(8) unsigned NOT NULL auto_increment,
   `maxnumrep` int(10) unsigned NOT NULL default '0',
   PRIMARY KEY  (`numeropost`)
 ) ;
@@ -227,7 +227,7 @@ INSERT INTO t1 (numeropost,maxnumrep) VALUES (40143,1),(43506,2);
 
 CREATE TABLE `t2` (
       `mot` varchar(30) NOT NULL default '',
-      `topic` mediumint(8) unsigned NOT NULL default '0',
+      `topic` int(8) unsigned NOT NULL default '0',
       `dt` date,
       `pseudo` varchar(35) NOT NULL default '',
        PRIMARY KEY  (`topic`)
